requirements_vanilla.txt
gradio
PyMuPDF
datasette
python-dotenv
instructor[google-generativeai]
lancedb
transformers
scikit-learn
docx2txt
datasets
nltk
openai
tantivy
pylance
sentence-transformers
numpy<2
pandas
opik
google-genai